+++ src/components.js
@@ -16,6 +16,13 @@ const noobj = Object.freeze({})
 const noarr = Object.freeze([])
 const noop = () => {}
 
+function getType(type, value) {
+  if (type.meta.kind === 'union') {
+    return type.dispatch(value)
+  }
+  return type
+}
+
 function getOptions(options, defaultOptions, value) {
   if (t.Nil.is(options)) {
     return defaultOptions
@@ -52,7 +59,7 @@ function getFormComponent(type, options) {
   case 'subtype' :
     return getFormComponent(type.meta.type, options)
   default :
-    t.fail(`[${SOURCE}] unsupported type ${name}`)
+    t.fail(`[${SOURCE}] unsupported kind ${type.meta.kind} for type ${name}`)
   }
 }
 
@@ -558,9 +565,9 @@ export class Struct extends Component {
 
     for (const prop in props) {
       if (props.hasOwnProperty(prop)) {
-        const propType = props[prop]
-        const fieldsOptions = options.fields || noobj
         const propValue = value[prop]
+        const propType = getType(props[prop], propValue)
+        const fieldsOptions = options.fields || noobj
         const propOptions = getOptions(fieldsOptions[prop], noobj, propValue)
         inputs[prop] = React.createElement(getFormComponent(propType, propOptions), {
           key: prop,
@@ -741,10 +748,10 @@ export class List extends Component {
     const config = this.getConfig()
     const templates = this.getTemplates()
     const value = this.state.value
-    const type = this.typeInfo.innerType.meta.type
     return value.map((itemValue, i) => {
+      const itemType = getType(this.typeInfo.innerType.meta.type, itemValue)
       const itemOptions = getOptions(options.item, noobj, itemValue)
-      const ItemComponent = getFormComponent(type, itemOptions)
+      const ItemComponent = getFormComponent(itemType, itemOptions)
       const buttons = []
       if (!options.disableRemove) {
         buttons.push({
@@ -770,7 +777,7 @@ export class List extends Component {
       return {
         input: React.createElement(ItemComponent, {
           ref: i,
-          type,
+          type: itemType,
           options: itemOptions,
           value: itemValue,
           onChange: this.onItemChange.bind(this, i),
@@ -829,17 +836,18 @@ export class Form extends React.Component {
   }
 
   render() {
-    const type = this.props.type
-    const value = this.props.value
-    const options = getOptions(this.props.options, noobj, value)
     const { i18n, templates } = Form
 
     if (process.env.NODE_ENV !== 'production') {
-      assert(t.isType(type), `[${SOURCE}] missing required prop type`)
-      assert(t.Object.is(options), `[${SOURCE}] prop options, if specified, must be an object or a function returning an object`)
+      assert(t.isType(this.props.type), `[${SOURCE}] missing required prop type`)
+      assert(t.maybe(t.Object).is(this.props.options) || t.Function.is(this.props.options), `[${SOURCE}] prop options, if specified, must be an object or a function returning an object`)
       assert(t.Object.is(templates), `[${SOURCE}] missing templates config`)
       assert(t.Object.is(i18n), `[${SOURCE}] missing i18n config`)
     }
+
+    const value = this.props.value
+    const type = getType(this.props.type, value)
+    const options = getOptions(this.props.options, noobj, value)
 
     // this is in the render method because I need this._reactInternalInstance
     const uidGenerator = this.getUIDGenerator()